+++ core/src/main/scala/dev/guillaumebogard/idb/api/Backend.scala
@@ -20,4 +20,3 @@ type Backend[F[_]] = dev.guillaumebogard.idb.internal.Backend[F]
 
 object Backend:
   def apply[F[_]](using b: Backend[F]): Backend[F] = b
-